+++ opentelemetry-appender-tracing/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## vNext
 
-- Add `OpenTelemetryTracingBridge::with_scope()` to configure the
-  OpenTelemetry `InstrumentationScope` used by the appender logger.
+- Add custom `InstrumentationScope` support via
+  `OpenTelemetryTracingBridge::with_scope()` and
+  `OpenTelemetryTracingBridge::builder_with_scope()`.
   [3415](https://github.com/open-telemetry/opentelemetry-rust/issues/3415)
 
 ## 0.32.0